Refactored connection header
