+++ source/locales/en/translation.json
@@ -3,7 +3,7 @@
     "plugged": "You're plugged into",
     "notPlugged": "You're not plugged in",
     "incomingConnection": "Incoming connection!",
-    "alpha": "Plug Alpha Release - {{version}}"
+    "alpha": "Alpha - {{version}}"
   },
   "copy": {
     "copyText": "Copy to clipboard",
